feat(widgets): move SafeLayout to ui widget library
