+++ packages/termui/example/03_responsive_dashboard.dart
@@ -8,115 +8,6 @@ import 'package:characters/characters.dart';
 import 'package:termui/terminal/terminal.dart' as term;
 import 'package:termui/ui/ui.dart';
 
-/// A layout widget that checks constraints and renders a fallback message
-/// if the terminal is below the minimum required width or height.
-class SafeLayout extends Widget {
-  final Widget child;
-  final int minWidth;
-  final int minHeight;
-
-  const SafeLayout({
-    required this.child,
-    this.minWidth = 40,
-    this.minHeight = 10,
-  });
-
-  @override
-  Element createElement() => SafeLayoutElement(this);
-
-  @override
-  int getIntrinsicHeight(int width) => child.getIntrinsicHeight(width);
-}
-
-class SafeLayoutElement extends Element {
-  Element? fallbackElement;
-  Element? childElement;
-  bool useFallback = false;
-
-  SafeLayoutElement(SafeLayout super.widget);
-
-  @override
-  void mount(Element? parent) {
-    super.mount(parent);
-    rebuild();
-  }
-
-  @override
-  void update(Widget newWidget) {
-    super.update(newWidget);
-    rebuild();
-  }
-
-  @override
-  void rebuild() {
-    final sl = widget as SafeLayout;
-    final fallback = Center(
-      child: Text(
-        'Screen too small!',
-        style: Style(foreground: Colors.red, modifiers: Modifier.bold),
-      ),
-    );
-
-    if (fallbackElement != null) {
-      fallbackElement!.update(fallback);
-    } else {
-      fallbackElement = fallback.createElement();
-      fallbackElement!.mount(this);
-    }
-
-    if (childElement != null &&
-        childElement!.widget.runtimeType == sl.child.runtimeType) {
-      childElement!.update(sl.child);
-    } else {
-      childElement?.unmount();
-      childElement = sl.child.createElement();
-      childElement!.mount(this);
-    }
-  }
-
-  @override
-  void visitChildren(void Function(Element child) visitor) {
-    if (useFallback) {
-      if (fallbackElement != null) visitor(fallbackElement!);
-    } else {
-      if (childElement != null) visitor(childElement!);
-    }
-  }
-
-  @override
-  Size performLayout(BoxConstraints constraints) {
-    final sl = widget as SafeLayout;
-    final w = constraints.maxWidth == BoxConstraints.infinity
-        ? 0
-        : constraints.maxWidth;
-    final h = constraints.maxHeight == BoxConstraints.infinity
-        ? 0
-        : constraints.maxHeight;
-
-    if (w < sl.minWidth || h < sl.minHeight) {
-      useFallback = true;
-      if (fallbackElement != null) {
-        fallbackElement!.layout(constraints);
-      }
-    } else {
-      useFallback = false;
-      if (childElement != null) {
-        childElement!.layout(constraints);
-      }
-    }
-    return constraints.constrain(Size(w, h));
-  }
-
-  @override
-  void paint(Buffer buffer, Offset offset) {
-    if (useFallback) {
-      fallbackElement?.paint(buffer, offset);
-    } else {
-      childElement?.paint(buffer, offset);
-    }
-  }
-}
-
 /// A component that renders status indicators with custom styles and borders.
 class StatCard extends StatelessWidget {
   final String title;
