+++ packages/termui/lib/ui/widget_toolkit.dart
@@ -41,3 +41,4 @@ export 'widgets/prompt_runner.dart';
 export 'widgets/horizontal_radio_group.dart';
 export 'widgets/selection_controller.dart';
 export 'widgets/layout_builder.dart';
+export 'widgets/safe_layout.dart';
